# docs/openapi/identity-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mktdash-identity-service",
    "description": "Authentication and authorization for Marketing Dashboard. Never called directly from the internet — the gateway verifies JWKS and forwards.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "EdDSA (Ed25519) access token. Verify against /.well-known/jwks.json; the `pv` claim is a cache key to validate, never an authority."
      }
    },
...
  },
  "paths": {
...
    "/v1/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "summary": "Consume a verification code, completing sign-up",
        "tags": [
          "authentication"
        ],
        "description": "Verifies the address, establishes a session and issues the token pair. POST, never GET: link-scanning proxies (Outlook Safe Links, Gmail's fetcher, corporate scanners) will fetch a URL before the human does, and a GET that consumes a single-use code produces 'the link doesn't work' reports nobody can reproduce.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 254
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "verified"
                      ]
                    },
                    "redirectTo": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "fullName": {
                          "type": "string"
                        },
                        "emailVerified": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "fullName",
                        "emailVerified"
                      ],
                      "additionalProperties": false
                    },
                    "organization": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "tenancy": {
                          "type": "string",
                          "enum": [
                            "company",
                            "agency"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "slug",
                        "tenancy"
                      ],
                      "additionalProperties": false
                    },
                    "workspace": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "slug"
                      ],
                      "additionalProperties": false
                    },
                    "membership": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "roleId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "roleSlug": {
                          "type": "string"
                        },
                        "dataScope": {
                          "type": "string",
                          "enum": [
                            "own",
                            "team",
                            "workspace",
                            "organization"
                          ]
                        },
                        "permissionVersion": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "id",
                        "roleId",
                        "roleSlug",
                        "dataScope",
                        "permissionVersion"
                      ],
                      "additionalProperties": false
                    },
                    "tokens": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "description": "EdDSA (Ed25519) JWT. Send as a Bearer token."
                        },
                        "tokenType": {
                          "type": "string",
                          "enum": [
                            "Bearer"
                          ]
                        },
                        "expiresIn": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991,
                          "description": "Access token lifetime in seconds."
                        },
                        "refreshToken": {
                          "type": "string",
                          "description": "Opaque, 256-bit, single-use. Rotated on every refresh; replaying a used token kills the whole family."
                        },
                        "refreshExpiresIn": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "accessToken",
                        "tokenType",
                        "expiresIn",
                        "refreshToken",
                        "refreshExpiresIn"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "status",
                    "redirectTo",
                    "user",
                    "organization",
                    "workspace",
                    "membership",
                    "tokens"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "description": "Stable, machine-readable. Branch on this, never on `detail`."
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Field paths only — the submitted value is never echoed back.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "detail"
                  ],
                  "additionalProperties": false,
                  "description": "RFC 9457 Problem Details"
                }
              }
            }
          },
          "410": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "description": "Stable, machine-readable. Branch on this, never on `detail`."
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Field paths only — the submitted value is never echoed back.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "detail"
                  ],
                  "additionalProperties": false,
                  "description": "RFC 9457 Problem Details"
                }
              }
            }
          },
          "429": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "description": "Stable, machine-readable. Branch on this, never on `detail`."
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Field paths only — the submitted value is never echoed back.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "detail"
                  ],
                  "additionalProperties": false,
                  "description": "RFC 9457 Problem Details"
                }
              }
            }
          },
          "500": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "description": "Stable, machine-readable. Branch on this, never on `detail`."
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "description": "Field paths only — the submitted value is never echoed back.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "detail"
                  ],
                  "additionalProperties": false,
                  "description": "RFC 9457 Problem Details"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "This service, behind the gateway"
    }
  ],
...
}